handlers separation and test fix
